require fabric

+++ flake.nix
@@ -107,9 +107,14 @@
             set -e
             export PATH=${with pkgs; lib.makeSearchPath "bin" [bash jq gawk perl findutils coreutils]}
             cd ~/.local/share/PrismLauncher/instances/"''${1?}"/ >/dev/null
+            if ! `jq -r .components\|any\(.cachedName=='"Fabric Loader"'\) mmc-pack.json`; then
+              echo "no Fabric :(" >&2
+              exit 1
+            fi
             jq -r .components[]\|select\(.cachedName==\"Minecraft\"\).cachedVersion mmc-pack.json | xargs bash ${builtins.toFile "install-2-electric-boogaloo" ''
               set -e
               cd .minecraft/mods/ >/dev/null
+              test -f fabric-api-*.jar || echo "Remember to install Fabric API — proceeding anyway…"
               rm -f goofyfiguraplugin-*.jar
               pushd "$1"/fabric/build/libs/ >/dev/null
               ls goofyfiguraplugin-*+$2.jar | sort -h | head -1 | xargs -i cp {} ~1
